+++ utils/timehandler.h
@@ -15,7 +15,7 @@ public:
     bool configure(unsigned long sampleRate, unsigned long bufferSize);
     tick_t getCurrentTick() const;
     tick_t msToTicks(unsigned long time_ms) const;
-    tick_t segToTicks(unsigned long time_seg) const;
+    tick_t segToTicks(double time_seg) const;
     unsigned long ticksToMs(tick_t) const;
 private:
     TimeHandler();
